program/bayes/requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.10
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.2
django-discover-runner @ file:///D:/package_python_offline/django_role_permissions/django_discover_runner-1.0-py2.py3-none-any.whl
django-extensions @ file:///D:/package_python_offline/django_extensions/django_extensions-3.0.9-py3-none-any.whl
django-import-export==2.4.0
django-role-permissions @ file:///D:/package_python_offline/django_role_permissions/django-role-permissions-3.1.0.tar.gz
entrypoints==0.3
et-xmlfile==1.0.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.2.0
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
mixed-naive-bayes==0.0.1
mysqlclient==1.4.6
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
nltk==3.5
notebook==6.1.4
numpy==1.19.2+mkl
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
pycparser==2.20
pydataset==0.2.0
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
regex==2020.10.11
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.50.2
traitlets==5.0.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
xlwt==1.3.0
